novatour/backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.0.1
pydantic>=2.11.0
pydantic-settings>=2.7.0
httpx>=0.28.0

# AWS
boto3>=1.35.0
strands-agents[bidi,bidi-io]>=1.30.0

# Google APIs (Places, Routes, Gemini Search)
google-genai>=1.14.0

# Utils
orjson>=3.10.0

# E2E Testing
pytest>=8.0.0
pytest-asyncio>=0.25.0
google-genai>=1.14.0
numpy>=1.26.0
